# docs/swagger.yaml
basePath: /api/v1
definitions:
  lib.Page:
    properties:
      first:
        description: indicate first value
        example: true
        type: boolean
      items:
        type: object
      last:
        description: indicate last value
        example: false
        type: boolean
      max_page:
        description: maximum pages for current schema
        example: 9
        type: integer
      page:
        description: current page, start from zero
        example: 0
        type: integer
      size:
        description: size per page, default `10`
        example: 10
        type: integer
      total:
        description: total data without filters
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 10
        type: integer
      visible:
        description: current length
        example: 10
        type: integer
    type: object
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.Ticket:
    properties:
      additional:
        type: string
      date:
        format: date
        type: string
      is_activated:
        type: boolean
      is_printed:
        type: boolean
      location_name:
        type: string
      movie_name:
        type: string
      price:
        type: number
      reference_number:
        type: string
      seat_code:
        type: string
      show_schedule_id:
        type: string
      start_time:
        format: date-time
        type: string
      theater_name:
        type: string
      transaction_id:
        type: string
    type: object
  model.Transaction:
    properties:
      additional:
        type: string
      booking_code:
        type: string
      contact_detail:
        type: string
      contact_name:
        type: string
      invoice_no:
        type: string
      tickets:
        items:
          $ref: '#/definitions/model.Ticket'
        type: array
      total_discount:
        type: number
      total_fee:
        type: number
      total_paid:
        type: number
      total_price:
        type: number
      total_ticket_price:
        type: number
      transaction_status:
        type: string
      user_id:
        format: uuid
        type: string
    type: object
...
  model.User:
    properties:
      additional:
        type: string
      email:
        description: Email
        example: armadamuhammads@gmail.com
        type: string
      first_name:
        description: FirstName
        example: Armada
        type: string
      is_admin:
        description: IsAdmin
        type: boolean
      last_login:
        description: LastLogin
        format: date-time
        type: string
      last_name:
        description: LastName
        example: Muhammad Siswanto
        type: string
      phone_number:
        description: PhoneNumber
        example: "089678009400"
        type: string
      username:
        description: Username
        type: string
    type: object
  model.UserAPI:
    properties:
      email:
        description: Email
        example: armadamuhammads@gmail.com
        type: string
      first_name:
        description: FirstName
        example: Armada
        type: string
      is_admin:
        description: IsAdmin
        type: boolean
      last_login:
        description: LastLogin
        format: date-time
        type: string
      last_name:
        description: LastName
        example: Muhammad Siswanto
        type: string
      phone_number:
        description: PhoneNumber
        example: "089678009400"
        type: string
      username:
        description: Username
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: armadamuhammads@gmail.com
    name: Armada Muhammad Siswanto
  description: API Documentation
  title: MKP Cinema Ticketing
  version: 1.0.0
paths:
...
  /transactions/{cart_id}:
    post:
      consumes:
      - application/json
      description: Post new Transaction
      parameters:
      - description: Cart ID
        in: path
        name: cart_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Post new Transaction
      tags:
      - Transaction
  /transactions/{id}:
    get:
      consumes:
      - application/json
      description: Get a Transaction by id
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Transaction by id
      tags:
      - Transaction
...
  /users:
    get:
      consumes:
      - application/json
      description: List of User
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of User
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of User
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create new User
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserAPI'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new User
      tags:
      - User
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get a User by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a User by id
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserAPI'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update User by id
      tags:
      - User
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: x-Token
    type: apiKey
swagger: "2.0"